requirements.txt
flask
gunicorn
pytesseract
pillow
pymupdf
reportlab